Added rviz functionality to the waypoint button being clicked, could make it a util function to use it across different buttons
